Finished first draft of enlarged memory map!

+++ src/main/kotlin/de/dude/library/repository/database/entity/Entity.kt
@@ -5,6 +5,6 @@ import de.dude.library.repository.database.Mapper
 abstract class Entity {
 
     @Suppress("LeakingThis")
-    val id: Int = Mapper.add(this)
+    val id: Long = Mapper.add(this)
 
 }
